added and moved files around to make a python package out of this repo
