+++ linear_planar/eddy_nlgc_mp.py
