+++ linear_planar/gs.py
