+++ linear_planar/smt_lin_pla.py
