+++ linear_planar/split_shells_and_types.py
